TEST: Extract type test for SimpleOlog
