+++ packages/documents/test/type-errors.test-d.ts
@@ -1,5 +1,4 @@
 import { PetriNet, Place, Transition } from "catcolab-logics/petri-net";
-import { Aspect, SimpleOlog, Type } from "catcolab-logics/simple-olog";
 import { AttrType, Mapping, SimpleSchema } from "catcolab-logics/simple-schema";
 import { describe, test } from "vitest";
 
@@ -12,25 +11,6 @@ import { describe, test } from "vitest";
 import { createBinder } from "catcolab-documents";
 
 describe.skip("type errors", () => {
-    test("invalid shapes are type errors in a simple olog", async () => {
-        const binder = createBinder();
-        const notebook = await binder.createNotebook(SimpleOlog, { title: "An Olog" });
-
-        const source = notebook.add(Type, { label: "A" });
-        const target = notebook.add(Type, { label: "B" });
-        const arrow = notebook.add(Aspect, { label: "has", from: source, to: target });
-
-        // @ts-expect-error Arrays are not valid endpoints in a simple olog.
-        arrow.update({ from: [source] });
-
-        // @ts-expect-error Arrays are not valid endpoints in a simple olog.
-        notebook.add(Aspect, { label: "bad", from: [source, target], to: target });
-        // @ts-expect-error Missing required fields.
-        notebook.add(Aspect, {});
-        // null fields are allowed.
-        notebook.add(Aspect, { label: null, from: null, to: null });
-    });
-
     test("a mapping's endpoints must be entities, not attribute types", async () => {
         const binder = createBinder();
         const schema = await binder.createNotebook(SimpleSchema, { title: "Example schema" });
