+++ packages/documents/tsconfig.test.json
@@ -7,5 +7,10 @@
         "jsxImportSource": "solid-js",
         "lib": ["ES2023", "DOM", "DOM.Iterable"]
     },
-    "include": ["vite.config.ts", "vitest.db-dump.config.ts", "test/**/*.ts", "test/**/*.tsx"]
+    "include": [
+        "vite.config.ts",
+        "vitest.db-dump.config.ts",
+        "test/creating-and-editing.test.ts",
+        "test/creating-and-editing.test-d.ts"
+    ]
 }
